chore: move studentSchema to global schema folder
