+++ features/manage-list/components/StudentFormDrawer/AcademicSection.tsx
@@ -2,7 +2,7 @@ import { useFormContext } from "react-hook-form";
 import FormSelect from "@/globals/components/shared/FormSelect";
 import { FieldGroup, FieldSet } from "@/globals/components/shad-cn/field";
 import SectionHeading from "./SectionHeading";
-import { StudentFormValues } from "@/features/auth/schema/studentSchema";
+import { StudentFormValues } from "@/globals/schemas/studentSchema";
 import { Option } from "@/globals/types/primitives";
 import { SchoolLevel, YearLevel } from "@prisma/client";
 import { useEffect } from "react";
