+++ features/manage-list/components/StudentFormDrawer/PersonalInfoSection.tsx
@@ -2,7 +2,7 @@ import SectionHeading from "./SectionHeading";
 import FormInput from "@/globals/components/shared/FormInput";
 import { FieldGroup, FieldSet } from "@/globals/components/shad-cn/field";
 import { useFormContext } from "react-hook-form";
-import { StudentFormValues } from "@/features/auth/schema/studentSchema";
+import { StudentFormValues } from "@/globals/schemas/studentSchema";
 
 const PersonalInfoSection = () => {
   const {
